3.7 Inversion de dependencias en modelo y paso de codigo a vista
